+++ workbench/c/Install-grub2-i386-pc.c
@@ -533,7 +533,7 @@ BOOL isvalidFileSystem(struct Volume * volume, CONST_STRPTR device,
 	    CloseRootPartition(ph);
 	}
 	else
-	    printf("Error OpenRootPartition(%s,%lu)\n", device, unit);
+	    printf("Error OpenRootPartition(%s,%lu)\n", device, (long)unit);
 	CloseLibrary((struct Library *) PartitionBase);
     }
     else
@@ -634,7 +634,7 @@ BOOL isvalidPartition(CONST_STRPTR device, ULONG unit, LONG * pnum,
 			{
 			    printf
 				("partition %ld not found on device %s unit %lu\n",
-				 *pnum, device, unit);
+				 (long)*pnum, device, (long)unit);
 			}
 		    }
 		    else
@@ -669,7 +669,7 @@ BOOL isvalidPartition(CONST_STRPTR device, ULONG unit, LONG * pnum,
 	    CloseRootPartition(ph);
 	}
 	else
-	    printf("Error OpenRootPartition(%s,%lu)\n", device, unit);
+	    printf("Error OpenRootPartition(%s,%lu)\n", device, (long)unit);
 	CloseLibrary((struct Library *) PartitionBase);
     }
     else
@@ -790,12 +790,12 @@ BOOL writeBootIMG(STRPTR bootimgpath, struct Volume * bootimgvol, struct Volume 
 		        error = writeBlock(bootimgvol, 0, bootimgvol->blockbuffer, 512);
                 
 		        if (error)
-		            printf("WriteError %lu\n", error);
+		            printf("WriteError %lu\n", (long)error);
 		        else
 		            retval = TRUE;
 	        }
 	        else
-		        printf("WriteError %lu\n", error);
+		        printf("WriteError %lu\n", (long)error);
 	    }
 	    else
 	        printf("%s: Read Error\n", bootimgpath);
@@ -831,7 +831,7 @@ ULONG collectBlockListFFS(struct Volume *volume, ULONG block,  struct BlockNode 
     if (retval)
     {
         D(bug("[install] collectBlockListFFS: ERROR reading block (error: %ld\n", retval));
-		printf("ReadError %lu\n", retval);
+		printf("ReadError %lu\n", (long)retval);
 		return 0;
 	}
 
@@ -849,7 +849,7 @@ ULONG collectBlockListFFS(struct Volume *volume, ULONG block,  struct BlockNode 
 		if (retval)
 		{
             D(bug("[install] collectBlockListFFS: ERROR reading block (error: %ld)\n", retval));
-			printf("ReadError %lu\n", retval);
+			printf("ReadError %lu\n", (long)retval);
 			return 0;
 		}
         
@@ -956,7 +956,7 @@ ULONG collectBlockListSFS(struct Volume *volume, ULONG objectnode,  struct Block
     if (retval)
     {
         D(bug("[install] collectBlockListSFS: ERROR reading root block (error: %ld)\n", retval));
-		printf("ReadError %lu\n", retval);
+		printf("ReadError %lu\n", (long)retval);
 		return 0;
 	}
 
@@ -1019,7 +1019,7 @@ ULONG collectBlockListSFS(struct Volume *volume, ULONG objectnode,  struct Block
         if (retval)
         {
             D(bug("[install] collectBlockListSFS: ERROR reading block (error: %ld)\n", retval));
-		    printf("ReadError %lu\n", retval);
+		    printf("ReadError %lu\n", (long)retval);
 		    return 0;
 	    }
 
@@ -1445,7 +1445,7 @@ int main(int argc, char **argv)
 	        else
 	        {
 		        printf("%s is not on device %s unit %ld\n",
-		           grubpath, bootDevice, unit);
+		           grubpath, bootDevice, (long)unit);
 		        ret = RETURN_ERROR;
 	        }
 	    }
