+++ workbench/c/install-i386-pc.c
@@ -469,7 +469,7 @@ D(bug("[install-i386] isvalidFileSystem(%x, %s, %d)\n", volume, device, unit));
 			CloseRootPartition(ph);
 		}
 		else
-			printf("Error OpenRootPartition(%s,%lu)\n", device, unit);
+			printf("Error OpenRootPartition(%s,%lu)\n", device, (long)unit);
 		CloseLibrary((struct Library *)PartitionBase);
 	}
 	else
@@ -576,7 +576,7 @@ D(bug("[install-i386] isvalidPartition(%s:%d, part:%d)\n", device, unit, pnum));
 							printf
 							(
 								"partition %ld not found on device %s unit %lu\n",
-								*pnum, device, unit
+								(long)*pnum, device, (long)unit
 							);
 						}
 					}
@@ -610,7 +610,7 @@ D(bug("[install-i386] isvalidPartition(%s:%d, part:%d)\n", device, unit, pnum));
 			CloseRootPartition(ph);		
 		}
 		else
-			printf("Error OpenRootPartition(%s,%lu)\n", device, unit);
+			printf("Error OpenRootPartition(%s,%lu)\n", device, (long)unit);
 		CloseLibrary((struct Library *)PartitionBase);
 	}
 	else
@@ -677,7 +677,7 @@ D(bug("[install-i386] collectBlockList: Cleared sector list (20 entries) [start:
    if (retval)
 	{
 D(bug("[install-i386] collectBlockList: ERROR reading block (error: %ld\n", retval));
-			printf("ReadError %lu\n", retval);
+			printf("ReadError %lu\n", (long)retval);
 			return 0;
 		}
 
@@ -698,7 +698,7 @@ D(bug("[install-i386] collectBlockList: First block @ %x, i:%d\n", first_block, 
 		if (retval)
 		{
 D(bug("[install-i386] collectBlockList: ERROR reading block (error: %ld)\n", retval));
-			printf("ReadError %lu\n", retval);
+			printf("ReadError %lu\n", (long)retval);
 			return 0;
 		}
 D(bug("[install-i386] collectBlockList: read block %lx, i = %d\n", block, i));
@@ -1005,12 +1005,12 @@ D(bug("[install-i386] writeStage1: Install to FLOPPY\n"));
 				error = readwriteBlock
 					(volume, 0, volume->blockbuffer, 512, volume->writecmd);
 				if (error)
-					printf("WriteError %lu\n", error);
+					printf("WriteError %lu\n", (long)error);
 				else
 					retval = TRUE;
 			}
 			else
-				printf("WriteError %lu\n", error);
+				printf("WriteError %lu\n", (long)error);
 		}
 		else
 			printf("%s: Read Error\n", stage1path);
@@ -1128,7 +1128,7 @@ D(bug("[install-i386] FORCELBA = %d\n",myargs[4]));
 								);
 						}
 						else
-							printf("Read Error: %lu\n", retval);
+							printf("Read Error: %lu\n", (long)retval);
 						uninitVolume(bbvol);
 					}
 					uninitVolume(grubvol);
@@ -1139,7 +1139,7 @@ D(bug("[install-i386] FORCELBA = %d\n",myargs[4]));
 				printf
 				(
 					"%s is not on device %s unit %lu\n",
-					(STRPTR)myargs[3], (STRPTR)myargs[0], *((ULONG *)myargs[1])
+					(STRPTR)myargs[3], (STRPTR)myargs[0], (long)*((ULONG *)myargs[1])
 				);
 			}
 		}
